refactor: ChatPageのコアロジックをuseChatフックに抽出 (#352)
- 12個のuseState・5個のuseEffect・2個のuseRefをuseChatに移動
- メッセージ管理・WebSocket接続・選択モード・言い換え提案を統合
- ChatPage: 489→207行に削減（58%削減）
- useChatテスト10件追加
